wip: jack playback, debug process optimization (x3)

+++ src/audio/Mixer.h
@@ -48,7 +48,7 @@ public:
 
   [[nodiscard]]
     int getEffectsChannelsCount() const {
-      return effectsChannels.size();
+      return static_cast<int>(effectsChannels.size());
     }
 
     bool addEffectsChannel();
